refactor: extract DXF block leaf entity emitters
