+++ plugins/CMakeLists.txt
@@ -13,6 +13,7 @@ set_target_properties(cadgf_sample_plugin PROPERTIES
 add_library(cadgf_dxf_importer_plugin SHARED
     dxf_importer_plugin.cpp
     dxf_block_entry_committers.cpp
+    dxf_block_leaf_entity_emitters.cpp
     dxf_block_header.cpp
     dxf_parser_helpers.cpp
     dxf_parser_zero_record.cpp
